requirements.txt
blinker
click
colorama
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
packaging
psycopg2
Werkzeug
python-decouple